Feature/review member

+++ src/main/java/com/puremadeleine/viewith/service/ReviewService.java
@@ -1,15 +1,14 @@
 package com.puremadeleine.viewith.service;
 
+import com.puremadeleine.viewith.domain.member.MemberEntity;
 import com.puremadeleine.viewith.domain.review.ReviewEntity;
 import com.puremadeleine.viewith.domain.review.ReviewReportEntity;
 import com.puremadeleine.viewith.domain.venue.SeatEntity;
 import com.puremadeleine.viewith.domain.venue.VenueEntity;
 import com.puremadeleine.viewith.dto.common.SortType;
 import com.puremadeleine.viewith.dto.review.*;
-import com.puremadeleine.viewith.provider.ReviewProvider;
-import com.puremadeleine.viewith.provider.ReviewReportProvider;
-import com.puremadeleine.viewith.provider.SeatProvider;
-import com.puremadeleine.viewith.provider.VenueProvider;
+import com.puremadeleine.viewith.exception.ViewithException;
+import com.puremadeleine.viewith.provider.*;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
@@ -20,6 +19,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static com.puremadeleine.viewith.converter.review.ReviewServiceConverter.toReviewListResDto;
+import static com.puremadeleine.viewith.exception.ViewithErrorCode.PERMISSION_DENIED_FOR_REVIEW;
 
 @Service
 @RequiredArgsConstructor
@@ -30,26 +30,31 @@ public class ReviewService {
     VenueProvider venueProvider;
     SeatProvider seatProvider;
     ReviewReportProvider reviewReportProvider;
+    MemberProvider memberProvider;
     ReviewServiceMapper mapper;
 
     @Transactional
-    public CreateReviewResDto createReview(CreateReviewReqDto reqDto) {
+    public CreateReviewResDto createReview(CreateReviewReqDto reqDto, Long memberId) {
+        MemberEntity activeMember = memberProvider.getActiveMember(memberId);
         VenueEntity venue = venueProvider.getVenue(reqDto.getVenueId());
         SeatEntity seat = seatProvider.getSeat(reqDto.getSection(), reqDto.getSeatRow(), reqDto.getSeatColumn());
-        ReviewEntity review = ReviewEntity.createReview(reqDto, venue, seat);
+
+        ReviewEntity review = ReviewEntity.createReview(reqDto, venue, seat, activeMember);
         ReviewEntity savedReview = reviewProvider.saveReview(review);
         return CreateReviewResDto.builder().reviewId(savedReview.getId()).build();
     }
 
     @Transactional
-    public void updateReview(Long reviewId, UpdateReviewReqDto reqDto) {
+    public void updateReview(Long reviewId, UpdateReviewReqDto reqDto, Long memberId) {
         ReviewEntity review = reviewProvider.getNormalReview(reviewId);
+        checkPermission(review.getMember().getId(), memberId);
         review.updateReview(reqDto);
     }
 
     @Transactional
-    public void deleteReview(Long reviewId) {
+    public void deleteReview(Long reviewId, Long memberId) {
         ReviewEntity review = reviewProvider.getNormalReview(reviewId);
+        checkPermission(review.getMember().getId(), memberId);
         review.deleteReview();
     }
 
@@ -65,18 +70,29 @@ public class ReviewService {
         return toReviewListResDto(isSummary, reviewList);
     }
 
-    public void reportReview(Long reviewId, ReportReviewReqDto req) {
+    public void reportReview(Long reviewId, ReportReviewReqDto req, Long memberId) {
         ReviewEntity review = reviewProvider.getNormalReview(reviewId);
+        if (review.getMember().getId().equals(memberId)) throw new ViewithException(PERMISSION_DENIED_FOR_REVIEW);
         review.reportReview();
         ReviewReportEntity reviewReport =
                 ReviewReportEntity.createReviewReport(review, req.getReportReason(), req.getReportReasonDetail());
         reviewReportProvider.saveReviewReport(reviewReport);
     }
 
+
+    private void checkPermission(Long reviewWriterId, Long reqMemberId) {
+        if (!reviewWriterId.equals(reqMemberId)) {
+            throw new ViewithException(PERMISSION_DENIED_FOR_REVIEW);
+        }
+    }
+
     @Mapper(componentModel = "spring")
     public interface ReviewServiceMapper {
 
         @Mapping(source = "id", target = "reviewId")
+        @Mapping(source = "member", target = "userInfo")
+        @Mapping(source = "member.id", target = "userInfo.userId")
+        @Mapping(source = "member.nickname", target = "userInfo.userNickname")
         ReviewInfoResDto toReviewInfoResDto(ReviewEntity review);
     }
